requirements.txt
fastapi
uvicorn
jinja2
pandas
langchain
openai
faiss-cpu
tiktoken
python-multipart
python-dotenv
sentence-transformers
langchain
langchain-community
langchain_groq
langchain-text-splitters
langchain-huggingface
uuid
openpyxl
